refactor: relocate NewEpochState decoders into amaru-kernel
